server/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
firebase-admin==6.0.1
google-cloud-storage==3.4.0
werkzeug==3.0.1
python-dotenv==1.0.0
